paralleldots/requirements.txt
Pillow
pandas
request
Flask
Flask-Cors
easyocr
torch===1.7.1
torchvision===0.8.2
torchaudio===0.7.2